test custom tool
